feat(config): Add initial configuration files and modules
This commit introduces several configuration files and modules for the
Atlantis-Nix project. It includes a new `.envrc` for shell configuration,
`.gitignore` to exclude specific files, and `clan.nix` for defining
machine inventory and services. Additionally, it adds `flake.lock` and
`flake.nix` for managing dependencies, along with a GNOME module for
enabling the GNOME desktop environment. A key configuration file for
SOPS is also included to manage encryption keys.